src/golang/contest/week198/5464.换酒问题_test.go
package week198

import "testing"

func Test_numWaterBottles(t *testing.T) {
	type args struct {
		numBottles  int
		numExchange int
	}
	tests := []struct {
		name string
		args args
		want int
	}{
		// Case 1
		{
			name: "Case 1",
			args: args{
				numBottles:  9,
				numExchange: 3,
			},
			want: 13,
		},
		// Case 2
		{
			name: "Case 2",
			args: args{
				numBottles:  15,
				numExchange: 4,
			},
			want: 19,
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := numWaterBottles(tt.args.numBottles, tt.args.numExchange); got != tt.want {
				t.Errorf("numWaterBottles() = %v, want %v", got, tt.want)
			}
		})
	}
}
